+++ app/services/permitted_attributes/address.rb
@@ -1,0 +1,11 @@
+module PermittedAttributes
+  class Address
+    def self.attributes
+      [
+        :firstname, :lastname, :address1, :address2,
+        :city, :country_id, :state_id, :zipcode,
+        :phone, :state_name, :alternative_phone, :company
+      ]
+    end
+  end
+end
